Move CPLB into its own package
