+++ pkg/component/controller/cplb/cplb_other.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package cplb
 
 import (
 	"context"
