+++ pkg/component/controller/cplb/cplb_reconciler_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package cplb
 
 import (
 	"testing"
